Finished Scene 8
Only two more scenes left to go before the alpha demo can be released to the public!

The last multi-choice scene and the climax!

Only 20 more days until the self-imposed deadline. I am pretty confident that we can get the rest of this packaged up and delivered as alpha demo before then!.

Best of luck!